# apps/drive/public/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Xenode Internal API",
    "description": "🔒 **INTERNAL USE ONLY** — This API reference is for internal development and integration purposes. All endpoints require authentication via session cookie or API key bearer token.\n\nXenode is a zero-knowledge encrypted cloud storage platform. The backend never sees raw file contents or unencrypted keys. All E2EE operations (encrypt, decrypt, key wrapping) are performed client-side.",
    "version": "0.1.0",
    "contact": {
      "name": "Xenode Engineering"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current environment"
    }
  ],
...
  "security": [
    { "bearerAuth": [] },
    { "cookieAuth": [] }
  ],
  "paths": {
...
    "/api/objects/{id}": {
      "get": {
        "tags": ["Objects"],
        "summary": "Get file download URL",
        "description": "Returns pre-signed download URL (or chunk URLs for chunked files) with encryption metadata.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Download details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": { "type": "string", "format": "uri" },
                    "chunkUrls": { "type": "array", "items": { "type": "string", "format": "uri" } },
                    "isEncrypted": { "type": "boolean" },
                    "encryptedDEK": { "type": "string" },
                    "iv": { "type": "string" },
                    "encryptedName": { "type": "string" },
                    "contentType": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": { "description": "Object not found" }
        }
      },
      "patch": {
        "tags": ["Objects"],
        "summary": "Update object metadata (tags, position)",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tags": { "type": "array", "items": { "type": "string" }, "example": ["work", "important"] },
                  "position": { "type": "integer", "example": 1 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Object updated" },
          "404": { "description": "Not found" }
        }
      },
      "delete": {
        "tags": ["Objects"],
        "summary": "Delete a file or folder",
        "description": "Moves the object into the Bin. Physical deletion and quota release occur through the permanent-deletion lifecycle.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Object deleted" },
          "404": { "description": "Not found" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "adminCookieAuth": {
        "type": "apiKey", "in": "cookie", "name": "__Host-Xenode_admin_session",
        "description": "Drive Admin JWT cookie, validated against the current Admin record and session version. Development uses Xenode_admin_session. This is separate from user authentication."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key authentication. Generate keys via `POST /api/keys`."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Session cookie managed by better-auth. 7-day expiry, auto-refreshes every 24h."
      },
      "cronAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Cron job secret token (CRON_SECRET env var)."
      }
    },
...
  }
}